+++ pages/enrollment/success.js
@@ -17,7 +17,7 @@ export default function success() {
       <SectionTitle name="Thank You!" />
       <div class="container">
         <div class="content is-large has-text-centered has-text-justified">
-          <p class="is-size-3">
+          <p class="is-size-3-desktop is-size-4">
             Your details have been successfully recorded. Lorem ipsum dolor sit
             amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
             ut labore et dolore magna aliqua.